% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.corregp}
\alias{print.corregp}
\title{Printing Correspondence Regression}
\usage{
\method{print}{corregp}(x, nf = 2, ...)
}
\arguments{
\item{x}{The output of a call to \code{\link{corregp}} (i.e. an object of class "corregp").}

\item{nf}{The number of dimensions to print. Defaults to the first two dimensions.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The output of a call to \code{\link{corregp}}.
}
\description{
Method to print the output of \code{\link{corregp}}.
}
\examples{
data(HairEye)
haireye.crg <- corregp(Eye ~ Hair * Sex, data = HairEye, b = 3000)
haireye.crg
print(haireye.crg, nf = 3)
}
\seealso{
\code{\link{corregp}}.
}

